crud after last update

+++ src/main/java/tn/esprit/pi/services/AdsService.java
@@ -1,6 +1,7 @@
 package tn.esprit.pi.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +16,7 @@ public class AdsService implements IAdsService {
 	@Override
 	public List<Ads> retrieveAds() {
 		
-		return (List<Ads>) adsRepository.findAll();
+		return  adsRepository.findAll();
 	}
 
 	@Override
@@ -31,12 +32,27 @@ public class AdsService implements IAdsService {
 
 	@Override
 	public Ads updateAds(Ads ad) {
-		return adsRepository.save(ad);
+		Ads ads = adsRepository.findById(ad.getIdAds()).get();
+		ads.setNameAds(ad.getNameAds());
+		ads.setChannel(ad.getChannel());
+		ads.setStart_date(ad.getStart_date());
+		ads.setFinal_date(ad.getFinal_date());
+		ads.setNb_initial_views(ad.getNb_initial_views());
+		ads.setNb_final_views(ad.getNb_final_views());
+		ads.setCost(ad.getCost());
+		ads.setType_ads(ad.getType_ads());
+		ads.setProducts(ad.getProducts());
+		return adsRepository.save(ads);
 	}
 
 	@Override
 	public void deleteAds(Long idAds) {
+		Optional<Ads> ads = adsRepository.findById(idAds);
+		if(ads.isPresent()){
 		adsRepository.deleteById(idAds);
+		}else{
+			System.out.println("No Ad record exist for given id");
+		}
 		
 	}
