+++ src/main/java/tn/esprit/pi/services/ProductCategoryService.java
@@ -1,12 +1,13 @@
 package tn.esprit.pi.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import tn.esprit.pi.entities.ProductCategory;
-
+import tn.esprit.pi.entities.Products;
 import tn.esprit.pi.repository.ProductCategoryRepository;
 @Service
 public class ProductCategoryService implements IProductCategoryService {
@@ -32,13 +33,21 @@ public class ProductCategoryService implements IProductCategoryService {
 
 	@Override
 	public ProductCategory updateProductCategory(ProductCategory pc) {
+		ProductCategory productCategory = productCategoryRepository.findById(pc.getIdCategoryProd()).get();
+		productCategory.setName_CategoryProd(pc.getName_CategoryProd());
+		productCategory.setProducts(pc.getProducts());
 		return productCategoryRepository.save(pc);
 	}
 
 	@Override
 	public void deleteProductCategory(Long idCategoryProd) {
-		productCategoryRepository.deleteById(idCategoryProd);
-		
+		Optional<ProductCategory> productcategory = productCategoryRepository.findById(idCategoryProd);
+		if(productcategory.isPresent()){
+			productCategoryRepository.deleteById(idCategoryProd);
+		}else{
+			System.out.println("No Product Category record exist for given id");
+		}
+				
 	}
 
 }
